Refactor: Move coverage code into a new coverage module
