+++ src/utils/fetch.js
@@ -1,0 +1,6 @@
+export const JSON_HEADERS = {
+  Accept: 'application/json',
+};
+export const PLAIN_HEADERS = {
+  Accept: 'text/plain',
+};
